Worked on Signal code - IN PROGRESS
Still making progress on the Signal code. The Test_Signal has not been
verified yet. I added a variable TVO to keep track of overflows and give
better precision, but I'm not sure if it works yet.

+++ Signal.h
@@ -15,8 +15,8 @@ unsigned int Signal_GetADC(unsigned char);
 
 float Signal_Voltage(unsigned char);
 void Signal_GetPhase(float);
-float Signal_GetPeriod();
-float Signal_GetFrequency();
+double Signal_GetPeriod();
+double Signal_GetFrequency();
 
 void Signal_Wait1s();
